Move unauthorized() out of getDatasetHelpers and into utils
It's used more widely than just by getDataset() and more generic than
even just the Charon endpoints.  I'll be renaming the getDatasetHelpers
module itself shortly.
